requirements.txt
# Core dependencies
python-dotenv>=1.0.0
openai>=1.0.0
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
